revice hieralchical structure
