+++ skstan/regression/linear_models/linear_models.py
@@ -1,40 +1,5 @@
-import numpy as np
-import pystan as ps
-
-from ..base import BaseModel, BaseStanData
-
-
-class RegressionStanData(BaseStanData):
-    def __init__(self, x: np.array, y: np.array, shrinkage: float):
-        super().__init__()
-        assert len(y.shape) == 1
-        assert len(x.shape) == 2
-        assert shrinkage >= 0
-
-        self.data = {
-            'x': x,
-            'y': y,
-            'n': x.shape[0],
-            'f': x.shape[1],
-            'shrinkage': shrinkage,
-        }
-
-
-class RegressionModelMixin(BaseModel):
-    def __init__(self, shrinkage: float):
-        self.shrinkage = shrinkage
-
-    def fit(self, x: np.array, y: np.array):
-        return ps.stan(
-            model_code=self.model_code,
-            data=self.preprocess(
-                RegressionStanData(x, y, self.shrinkage)
-            ).data
-        )
-
-    @staticmethod
-    def preprocess(dat: RegressionStanData) -> RegressionStanData:
-        return dat
+from ..base import RegressionModelMixin
+from ..base import RegressionStanData
 
 
 class LinearRegression(RegressionModelMixin):
